requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
polars
google-generativeai
pyarrow
google-cloud-storage
google-cloud-bigquery
astronomer-cosmos
